+++ README.md
@@ -14,14 +14,16 @@ source env/bin/activate
 pip install -r requirements.txt
 ```
 
-Run `pascal2csharp` module with
+Run `transpiler` module with
 
 ```
-python -m pascal2csharp
+python -m transpiler path/to/source/code
+e.g.
+python -m transpiler examples/supported_syntax.pas
 ```
 
 Before submitting a pull request make sure that your code passes
-all tests and there is no `flake8` linter errors. Check it with
+all tests and there are no `flake8` linter errors. Check it with
 
 ```
 python -m unittest
